Remove commented-out leftovers from older versions
Version 2019.195 claimed that a subtraction operator would be too
"complicated" and has therefore been left out deliberately.

Although I am not sure whether this is actually true at this point, I
give in to that claim and remove the subtraction operator which was
already present in an earlier version.

I also remove the addition operator, because the claiming version did
the same without explaining why.

Anyway, the old version can still be restored in case I come to the
conclusion later that there is no harm in providing an addition or
subtraction operator.

BTW, as I recall very dimly, the reason for the "complications" had to
do with the alignment. It would work in the wrong direction when
subtraction, or something like that.