Make scripts executable
This has not been so before because the scripts were also intended to be
used on filesystems which do not provide explicit execute permissions.

However, the scripts can still be used there; they use "sh" internally
for invoking one another, so they do not depend on executability in
order to function.

Also remove the .sh extension from some of the scripts, because either
all the scripts shall have them or none.